Finish offset extraction and add unit tests.
This commit completes the offset extraction feature. The example below
describes best how it works:

    ze::event event{"foo", 42u, ze::record{"foo", true, 4.2}};
    expr.parse("@1,0 ~ /foo/");
    assert(expr.eval(event));

Closes #4.

+++ src/vast/detail/ast.h
@@ -97,7 +97,7 @@ struct tag_clause
 
 struct offset_clause
 {
-  std::vector<unsigned> offsets;
+  std::vector<size_t> offsets;
   clause_operator op;
   expression rhs;
 };
@@ -184,7 +184,7 @@ BOOST_FUSION_ADAPT_STRUCT(
 
   BOOST_FUSION_ADAPT_STRUCT(
     vast::detail::ast::offset_clause,
-    (std::vector<unsigned>, offsets)
+    (std::vector<size_t>, offsets)
     (vast::detail::ast::clause_operator, op)
     (vast::detail::ast::expression, rhs))
